make storage backends pluggable
major changes:

* export ExpiringStorage class, not instance
* refactor location.hash handling as separate storage backend
* introduce new id property used as storage key
* introduce new storage property used as factory for storage backend
* introduce new storageOptions property used as configuration object
  for storage backend
* remove cacheLifetime property (now available as cacheLifeTimeInMinutes
  on the storageOptions property)
* remove options property (useUrlFragment option now available as
  useUrlFragment on the storageOptions property)
* export storage backends

storage backends:

a storage backend must have a get() and a set(value) method. an options
object is passed to the storage factory function that includes all
options from the storageOptions object, an autoSelectTab method to
trigger a reload from the storage backend(s) and a key property that can
be used used for storage lookups.